+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 ### Features
+- hel-1076: [Bloc Capacité et Autorisations] EJ - Afficher les autorisations des ET MS
 
 ## [1.8.14](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.13...release/1.8.14) (2026-07-28)